+++ src/index.ts
@@ -1,8 +1,15 @@
 export * as EmojiPicker from "./components/emoji-picker";
 export { defaultEmojiDataResolver } from "./data/emoji";
 export { createEmojiDataCache } from "./data/emoji-data-cache";
-export { getEmojiDetails } from "./data/get-emoji-details";
+export {
+  type GetEmojiDetailsOptions,
+  getEmojiDetails,
+} from "./data/get-emoji-details";
 export { useActiveEmoji, useSkinTone } from "./hooks";
+export {
+  type UseEmojiDetailsOptions,
+  useEmojiDetails,
+} from "./hooks/use-emoji-details";
 export type {
   Category,
   Emoji,